♻️ refactor: BottomSheetCheckIcon <Icon /> 사용 도입 #387
